test(proxy): run precedence coverage in proxy shard
